refactor: reorganize folder structure into modules
