+++ src/modules/roles/domain/role.dto.js
@@ -7,7 +7,7 @@
  */
 
 const Joi = require("joi");
-const RoleTypeEnum = require("../../_shared/enum/roles.enum");
+const RoleTypeEnum = require("../../../_shared/enum/roles.enum");
 
 module.exports = Joi.object({
   name: Joi.string()
