+++ src/modules/roles/domain/role.swagger.js
