+++ src/modules/roles/infrastructure/roles.routers.js
@@ -3,14 +3,14 @@ const roleController = require("./roles.controller");
 const router = express.Router();
 const roleDto = require("../domain/role.dto");
 const updateRoleDto = require("../domain/roleUpdate.dto");
-const validateBodyDto = require("../../_shared/middlewares/validate/dtoBody.validate");
-const validateId = require("../../_shared/middlewares/validate/id.validate");
-const authenticationMiddleware = require("../../_shared/middlewares/authentication.middleware");
-const authorizationMiddleware = require("../../_shared/middlewares/authorization.middleware");
-const roleDeleteMiddleware = require("../../_shared/middlewares/roleDelete.middleware");
-const RoleTypeEnum = require("../../_shared/enum/roles.enum");
-const paginateDto = require("../../_shared/joi/paginateDto.joi");
-const validateQueryDto = require("../../_shared/middlewares/validate/dtoQuery.validate");
+const validateBodyDto = require("../../../_shared/middlewares/validate/dtoBody.validate");
+const validateId = require("../../../_shared/middlewares/validate/id.validate");
+const authenticationMiddleware = require("../../../_shared/middlewares/authentication.middleware");
+const authorizationMiddleware = require("../../../_shared/middlewares/authorization.middleware");
+const roleDeleteMiddleware = require("../../../_shared/middlewares/roleDelete.middleware");
+const RoleTypeEnum = require("../../../_shared/enum/roles.enum");
+const paginateDto = require("../../../_shared/joi/paginateDto.joi");
+const validateQueryDto = require("../../../_shared/middlewares/validate/dtoQuery.validate");
 
 const middleAccess = [
   RoleTypeEnum.ADMIN,
